start renaming files to jsx
