+++ maltd-frontend/.babelrc
@@ -10,6 +10,7 @@
       ],
       "only": [
         "./**/*.js",
+        "./**/*.jsx",
         "node_modules/jest-runtime"
       ]
     }
