+++ maltd-frontend/.storybook/main.js
@@ -1,5 +1,5 @@
 module.exports = {
-  stories: ["../src/**/**/**/*.stories.js"],
+  stories: ["../src/**/**/**/*.stories.jsx"],
   addons: [
     "@storybook/preset-create-react-app",
     "@storybook/addon-actions",
