+++ maltd-frontend/package.json
@@ -18,7 +18,7 @@
     "react-icons": "^3.9.0",
     "react-scripts": "3.4.0",
     "react-test-renderer": "^16.12.0",
-    "reactstrap": "^8.4.0"
+    "reactstrap": "^8.4.1"
   },
   "scripts": {
     "start": "react-scripts start",
