+++ maltd-frontend/src/components/base/BackIcon/BackIcon.stories.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { action } from "@storybook/addon-actions";
 import { storiesOf } from "@storybook/react";
 import "bootstrap/dist/css/bootstrap.css";
-import BackIcon from "./BackIcon";
+import BackIcon from "./BackIcon.jsx";
 
 const backIcon = {
   message: "Find another user"
