+++ maltd-frontend/src/components/base/Dropdown/Dropdown.stories.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { storiesOf } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 import "bootstrap/dist/css/bootstrap.css";
-import Dropdown from "./Dropdown";
+import Dropdown from "./Dropdown.jsx";
 
 const dropdown = {
   defaultTitle: "Choose a project",
