+++ maltd-frontend/src/components/base/GeneralButton/GeneralButton.stories.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { action } from "@storybook/addon-actions";
 import { storiesOf } from "@storybook/react";
 import "bootstrap/dist/css/bootstrap.css";
-import GeneralButton from "./GeneralButton";
+import GeneralButton from "./GeneralButton.jsx";
 
 const generalButton = {
   type: "submit",
