+++ maltd-frontend/src/components/base/InputField/InputField.stories.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { action } from "@storybook/addon-actions";
 import { storiesOf } from "@storybook/react";
 import "bootstrap/dist/css/bootstrap.css";
-import InputField from "./InputField";
+import InputField from "./InputField.jsx";
 
 const inputField = {
   type: "text",
