+++ maltd-frontend/src/components/base/InputField/InputField.test.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import InputField from "./InputField";
+import InputField from "./InputField.jsx";
 
 describe("Input Field", () => {
   test("Component renders as expected", () => {
