+++ maltd-frontend/src/components/base/ListElement/ListElement.test.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import renderer from "react-test-renderer";
 import Adapter from "enzyme-adapter-react-16";
 import Enzyme, { shallow } from "enzyme";
-import ListElement from "./ListElement";
+import ListElement from "./ListElement.jsx";
 import XIcon from "../XIcon/XIcon";
 
 Enzyme.configure({ adapter: new Adapter() });
