+++ maltd-frontend/src/components/composite/UserAccess/UserAccess.js
@@ -1,9 +1,9 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { Table } from "reactstrap";
-import ListElement from "../../base/ListElement/ListElement";
+import ListElement from "../../base/ListElement/ListElement.jsx";
 import PlusIcon from "../../base/PlusIcon/PlusIcon";
-import Dropdown from "../../base/Dropdown/Dropdown";
+import Dropdown from "../../base/Dropdown/Dropdown.jsx";
 import "./UserAccess.css";
 
 export default function UserAccess({
